requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.4; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
djangorestframework==3.15.2; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
